web_scrap_client/requirements.txt
google-play-scraper
python-dotenv
requests
beautifulsoup4
selenium
webdriver-manager
